simple-adk-agent/sync-to-action/requirements.txt
google-adk
google-genai
fastapi
uvicorn
pydantic
python-dotenv
